Update page layout
